policy, mapper, capver, hscontrol: remove dead code
