+++ hscontrol/capver/capver.go
@@ -3,7 +3,6 @@ package capver
 //go:generate go run ../../tools/capver/main.go
 
 import (
-	"slices"
 	"sort"
 	"strings"
 
@@ -30,13 +29,6 @@ func tailscaleVersSorted() []string {
 	sort.Strings(vers)
 
 	return vers
-}
-
-func capVersSorted() []tailcfg.CapabilityVersion {
-	capVers := xmaps.Keys(capVerToTailscaleVer)
-	slices.Sort(capVers)
-
-	return capVers
 }
 
 // TailscaleVersion returns the Tailscale version for the given [tailcfg.CapabilityVersion].
@@ -66,21 +58,6 @@ func CapabilityVersion(ver string) tailcfg.CapabilityVersion {
 	return 0
 }
 
-// TailscaleLatest returns the n latest Tailscale versions.
-func TailscaleLatest(n int) []string {
-	if n <= 0 {
-		return nil
-	}
-
-	tsSorted := tailscaleVersSorted()
-
-	if n > len(tsSorted) {
-		return tsSorted
-	}
-
-	return tsSorted[len(tsSorted)-n:]
-}
-
 // TailscaleLatestMajorMinor returns the n latest Tailscale versions (e.g. 1.80).
 func TailscaleLatestMajorMinor(n int, stripV bool) []string {
 	if n <= 0 {
@@ -106,19 +83,4 @@ func TailscaleLatestMajorMinor(n int, stripV bool) []string {
 	}
 
 	return majorSl[len(majorSl)-n:]
-}
-
-// CapVerLatest returns the n latest [tailcfg.CapabilityVersion] values.
-func CapVerLatest(n int) []tailcfg.CapabilityVersion {
-	if n <= 0 {
-		return nil
-	}
-
-	s := capVersSorted()
-
-	if n > len(s) {
-		return s
-	}
-
-	return s[len(s)-n:]
 }
